improve reactopya-server

+++ reactopya/reactopya_server/server/SessionManager.js
@@ -2,6 +2,7 @@ import fs from 'fs';
 import { spawn } from 'child_process';
 import { Session } from 'inspector';
 import path from 'path';
+import { SIGINT } from 'constants';
 
 // Manage the app sessions
 export default class SessionManager {
@@ -32,17 +33,21 @@ export default class SessionManager {
 
         // Create a new session
         let S = new AppSession(this._config, this._workingDir, this._appConfigByName[appName], query);
+        console.info(`Session started: ${appName} ${S.sessionId()}`);
+        this._sessions[S.sessionId()] = S;
         S.onMessageFromPython(async (msg) => {
             for (let handler of this._messageFromPythonHandlers) {
                 await handler(S.sessionId(), msg);
             }
         });
+        S.onStopped(async () => {
+            console.info(`Session stopped: ${appName} ${S.sessionId()}`);
+            delete this._sessions[S.sessionId()];
+        });
         if (!await S.start()) {
             // it didn't start somehow
             return null;
         }
-
-        this._sessions[S.sessionId()] = S;
 
         // Return the new session
         return S;
@@ -53,6 +58,12 @@ export default class SessionManager {
             return;
         }
         await this._sessions[sessionId].handleMessageFromJavaScript(message);
+    }
+    async stopSession(sessionId) {
+        if (!(sessionId in this._sessions)) {
+            return;
+        }
+        await this._sessions[sessionId].stop();
     }
 }
 
@@ -68,6 +79,8 @@ class AppSession {
         this._pythonProcess = null; // will be created on start()
         this._status = 'pending'; // waiting to start
         this._messageFromPythonHandlers = []; // message handlers - see onMessageFromPython
+        this._stoppedHandlers = [];
+        this._lastKeepAlive = new Date();
     }
     sessionId() {
         return this._sessionId;
@@ -92,10 +105,17 @@ class AppSession {
         await this._checkForMessagesFromPython();
         return true;
     }
+    async stop() {
+        await this._doStop();
+    }
     async handleMessageFromJavaScript(message) {
+        if (message.name === 'keepAlive') {
+            this._lastKeepAlive = new Date();
+            return;
+        }
         let fname = `${this._sessionDir}/${this._messageIndex}.msg-from-js`;
         this._messageIndex++;
-        
+
         await fs.promises.writeFile(fname + '.tmp', JSON.stringify(message), 'utf-8');
         await fs.promises.rename(fname + '.tmp', fname);
     }
@@ -103,6 +123,9 @@ class AppSession {
         // Note that handler must be an async function!
         // register the handler
         this._messageFromPythonHandlers.push(handler);
+    }
+    onStopped(handler) {
+        this._stoppedHandlers.push(handler);
     }
     status() {
         // return the status, pending, running, stopped, error
@@ -116,7 +139,7 @@ class AppSession {
                 let html = await fs.promises.readFile(this._sessionDir + '/index.html', 'utf-8');
                 return html;
             }
-            catch(err) {
+            catch (err) {
                 let elapsed = (new Date()) - timer;
                 if (elapsed > 15000) {
                     // it's been too long
@@ -129,11 +152,28 @@ class AppSession {
         }
     }
     async _checkForMessagesFromPython() {
+        if (this._pythonProcess.exited()) {
+            await this._doStop();
+            return;
+        }
         if (this._status !== 'running')
             return;
+        let elapsedSinceLastKeepAlive = (new Date()) - this._lastKeepAlive;
+        if (elapsedSinceLastKeepAlive > 60 * 1000) {
+            await this._doStop();
+            return;
+        }
         // look for message files (from python) in the session directory
         let messageFiles = [];
-        let files = await fs.promises.readdir(this._sessionDir);
+        let files;
+        try {
+            files = await fs.promises.readdir(this._sessionDir);
+        }
+        catch(err) {
+            console.warn('Closing session because unable to read session directory.');
+            await this._doStop();
+            return;
+        }
         for (let file of files) {
             if (file.endsWith('.msg-from-py')) {
                 messageFiles.push(file);
@@ -157,7 +197,7 @@ class AppSession {
         }
         if (!fs.existsSync(this._sessionDir)) {
             // the session directory does not exist, so we are done
-            this._status = 'stopped';
+            await this._doStop();
             return;
         }
         setTimeout(async () => {
@@ -166,10 +206,39 @@ class AppSession {
             await this._checkForMessagesFromPython();
         }, 500);
     }
+    async _doStop() {
+        if (this._status === 'stopped')
+            return;
+        this._status = 'stopped';
+        await this._pythonProcess.stop();
+        try {
+            deleteFolderRecursive(this._sessionDir);
+        }
+        catch(err) {
+            // maybe it was already gone.
+        }
+        for (let handler of this._stoppedHandlers) {
+            await handler();
+        }
+    }
 }
 
 function waitMsec(ms) {
     return new Promise(resolve => setTimeout(resolve, ms));
+}
+
+function deleteFolderRecursive(path) {
+    if (fs.existsSync(path)) {
+        fs.readdirSync(path).forEach((file, index) => {
+            const curPath = Path.join(path, file);
+            if (fs.lstatSync(curPath).isDirectory()) { // recurse
+                deleteFolderRecursive(curPath);
+            } else { // delete file
+                fs.unlinkSync(curPath);
+            }
+        });
+        fs.rmdirSync(path);
+    }
 }
 
 class PythonProcess {
@@ -181,6 +250,10 @@ class PythonProcess {
         this._sessionId = sessionId;
         this._query = query; // they query from the url
         this._process = null; // will create at start()
+        this._exited = false;
+    }
+    exited() {
+        return this._exited;
     }
     async start() {
         // the config path is needed to resolve the location of the py files
@@ -192,12 +265,16 @@ class PythonProcess {
         }
         let pyFilePath = path.resolve(configPath, this._appConfig.pyFile);
 
-        // spawn the process
-        // note: eventually we will pass the query parameters into this
-        let env = Object.create( process.env );
+        // spawn the process and pass the query parameters
+        let env = Object.create(process.env);
         env.REACTOPYA_SERVER_SESSION_DIR = this._sessionDir;
         env.REACTOPYA_SERVER_SESSION_ID = this._sessionId;
-        this._process = spawn('python', [pyFilePath], {env: env});
+        let args = [pyFilePath]
+        for (let key0 in this._query) {
+            args.push(`--${key0}`);
+            args.push(this._query[key0]);
+        }
+        this._process = spawn('python', args, { env: env });
         this._process.stderr.on('data', (data) => {
             // stderr from the process
             console.error('FROM PROCESS:', data.toString());
@@ -206,7 +283,24 @@ class PythonProcess {
             // stdout from the process
             console.info('FROM PROCESS:', data.toString());
         });
+        this._process.on('exit', (code) => {
+            this._exited = true;
+        });
         return true;
+    }
+    async stop() {
+        if (this._exited) return;
+        let signals = ['SIGINT', 'SIGINT', 'SIGTERM', 'SIGTERM', 'SIGKILL', 'SIGKILL']
+        for (let sig of signals) {
+            this._process.kill(sig);
+            await waitMsec(500);
+            if (this._exited) return;
+            await waitMsec(4500);
+            if (this._exited) return;
+        }
+        if (!this._exited) {
+            console.warn('Unable to stop process after trying signals.');
+        }
     }
 }
 
@@ -220,6 +314,6 @@ function makeRandomId(num_chars) {
     var text = "";
     var possible = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789";
     for (var i = 0; i < num_chars; i++)
-      text += possible.charAt(Math.floor(Math.random() * possible.length));
+        text += possible.charAt(Math.floor(Math.random() * possible.length));
     return text;
 }
